Move parser tests over for the new parser module
